Will do final tests tomorrow, but all seems to work. Yay for efficiency and options for plugin developers!

+++ protocol/src/main/java/net/md_5/bungee/protocol/packet/DefinedPacket.java
@@ -39,7 +39,7 @@ public abstract class DefinedPacket
     public void writeArray(byte[] b, ByteBuf buf)
     {
         // TODO: Check len - use Guava?
-        buf.writeByte( b.length );
+        buf.writeShort( b.length );
         buf.writeBytes( b );
     }
